menu: fix moving submenu indicator
the submenu is part of the flex layout of the menu item, so initially,
when it's not displayed yet, additional space is used for the flex
'gap'.

when it's shown, the position will be set to fixed anyway and it's no
longer part of the flex layout, removing the additional gap and
subsequently moving the remaining items in the menu item (i.e. the
submenu arrow).

to fix this, set the initial position to fixed alread, making it never
part of the flex layout.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>