improve light/dark mode handling
by not providing one css file per mode, but by controlling the mode
with a class on the root html element.

This simplifies the delivery of the css files since there is only one
per theme now instead of two.

Needs changes to the themeloader too.

By default (if there is no class on the root element, the light theme
will be chosen)

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>