input: show outline for fields with open pickers
in dropdown/selector/etc. fields, the input loses the focus when the
picker is open and has focus (which it normally gets when it opens) so
the outline vanishes. This results in flicker since the picker does not
open instantaiously.

To prevent that, show the outline always when the picker is open,
indicating that this field has the focus.

The dropdown needs to add the 'picker-open' class to the input element
for this to work.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>