input: fix min-height for input elements
we want the input element to be at least as tall as the line-height,
otherwise it can happen that it's smaller, e.g. when using a custom
render_value function for dropdowns, as those use a hidden input field
that does not have a default size.

with this, the minimum height for fields should be consistent

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>