makefile: rework and improve handling the Roboto Flex font files
The previous makefile had a few issues, like using "quotes" as if make
variables where shell variables, but they ain't, one just needs quotes
when actually passing something to an command, otherwise they can
"stack" and even cause errors if processed by make itself, like if
they are used as for a file target.

Besides that, an error like having forgotten to install the package
providing woff2_compress would cause some friction, one has to know
which package it's shipped with and after installing that one a rerun
of the target will re-download the TTF file, which is not necessary.
So split these into separate targets and output a error with a hint if
the woff2_compress executable cannot be found.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>