change vh/vw to dvh/dvw
using 100vh/vw as the viewport height/width works most of the time, but
on mobile this might be bigger than the viewable part, since the address
bar sometimes is added to that size, even if visible.

to always really show the visible part, use 'dvh/dvw' which are the
'dynamic viewport height/width' and always is the viewable part of the
screen.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>