buttons: rework fab menu
this reworks the fab menu to adhere more to material3, and switch to use
flex layout instead of calculating pixel offsets. This eliminates most
offset calculations for the menu.

Also introduce animations like material3 suggests.

We now remove the left and right directions as well as the center
alignments, as they almost never make sense layout-wise.
(We can still re-implement them when we need it, but they make
the css much more complicated)

For this to work properly, the fab menu code has to be updated as well,
for the now different class hierarchy.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>